deepsearch/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langgraph
langchain-openai
tavily-python
jinja2
pydantic
pydantic_settings
